pyproject.toml
[project]
name = "lucky-mbti"
version = "0.1.0"
description = ""
authors = [
    {name = "nadaa",email = "nadaa@ajou.ac.kr"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "openai (>=1.61.0,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
